Migrate monetary fields to BigDecimal and enforce 2-decimal scale (#58)
* Migrate monetary fields from double to BigDecimal

Currency amounts stored as double risk silent rounding drift, which
matters once ledger reconciliation depends on exact balances.

- Entities (GiftCard, GiftCardHold, LedgerEntry) and DTOs now use
  BigDecimal for balance/amount fields.
- Services use compareTo/add/subtract instead of primitive operators.
- New Flyway migration (V15) alters existing DOUBLE PRECISION columns
  to NUMERIC(19,2).
- Idempotency-key hashing normalizes amounts to a fixed 2-decimal
  scale before hashing, so equivalent amounts with different JSON
  representations still hit the same cache key.
- Updated unit and integration tests accordingly.

* Reject monetary amounts with more than 2 decimal places

Manual smoke testing found that a balance sent as 99.999 got silently
rounded to 100.00 by the NUMERIC(19,2) column, while the create
response still echoed back the unrounded 99.999 from the in-memory
entity - a client-visible mismatch with what was actually persisted.

- Add @Digits(integer=17, fraction=2) to balance/amount fields on
  GiftCardCreateRequest, RedemptionRequest and ReserveRequest, so
  over-precise input is rejected with 400 instead of silently rounded.
- Add GiftCardMoneyValidationIntegrationTest covering the new
  constraint end-to-end via the real controllers.
- Fix GiftCardServiceIntegrationTest: it never cleaned users/
  refresh_tokens between tests, which only became a problem once
  another test class in the suite started creating real logged-in
  users - leftover rows broke its merchant cleanup via FK violation.

* Normalize monetary amounts to scale 2 before use

@Digits already rejects more than 2 decimal places, but amounts were
still echoed back at whatever scale the client sent them in (e.g.
"amount":1 produced "deductedAmount":1, not 1.00). That's numerically
correct but doesn't visibly confirm which amount was actually applied.

- executeRedemptionSync, createGiftCard and reserve() now setScale(2,
  HALF_UP) the validated amount/balance immediately, before it's used
  for balance math, ledger entries, or the response - so every
  monetary value in a response consistently shows 2 decimals.
- Add a test proving deductedAmount/remainingBalance come back as
  "1.00"/"99.00" from a bare "amount":1, asserting on the raw JSON
  body since jsonPath's numeric comparison can't tell "1" from "1.00".

+++ src/main/java/com/finovago/p2p/dto/GiftCardResponse.java
@@ -1,5 +1,6 @@
 package com.finovago.p2p.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -10,7 +11,7 @@ public record GiftCardResponse(
     String giftCardCode,
 
     @Schema(description = "Current balance", example = "100.0")
-    double balance,
+    BigDecimal balance,
 
     @Schema(description = "Indicates if the gift card is active", example = "true")
     boolean active,
